Added stopwatch+wait bindings, plus little variable comparison fix, hardened existing http tests (#14)

+++ NatLaRestTest.Bindings.Interfaces/Assertions/INumericVariablePathAssertions.cs
@@ -7,12 +7,12 @@ public interface INumericVariableAssertions
     /// </summary>
     /// <param name="variableName">The variable name containing a numeric value.</param>
     /// <param name="value">The threshold value (exclusive).</param>
-    void NumericVariableIsGreaterThan(string variableName, int value);
+    void NumericVariableIsGreaterThan(string variableName, double value);
 
     /// <summary>
     ///     Then step: Asserts that the numeric value stored in the specified variable is less than the given value.
     /// </summary>
     /// <param name="variableName">The variable name containing a numeric value.</param>
     /// <param name="value">The threshold value (exclusive).</param>
-    void NumericVariableIsLessThan(string variableName, int value);
+    void NumericVariableIsLessThan(string variableName, double value);
 }
